Version 2 - changing variable to unsigned long int

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,8 +8,8 @@
   */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0;
-	unsigned int zor;
+	unsigned long int count = 0;
+	unsigned long int zor;
 
 	zor = n ^ m;
